refactor(zhongwen): single-home the metered Epicenter engine builder
The browser tab and the daemon both built the metered Epicenter engine
inline, the same way: wrap the session fetch for the AI-chat route, point
at `/api/ai/chat` on the base URL, send `{ model, systemPrompts }`. Two
copies of one wire shape, free to drift (a system-prompt or model tweak
had to land in both).

Extract `epicenterMeteredEngine(sessionFetch, baseURL): Engine` and have
each peer call it with its own fetch and base. It cannot live in the
dep-free contract (`zhongwen.ts`) because it pulls in `@epicenter/client`,
so it gets its own subpath, `@epicenter/zhongwen/engine`; the contract
stays browser-safe. `createAiChatFetch` runs inside the returned engine,
so building it is free and the wrap happens only when the engine is
actually reached (the daemon still gates it behind the metered opt-in).

mount.ts drops its `@epicenter/client` and `API_ROUTES` imports;
ConversationView drops the same plus its hand-wired fetch. zhongwen
typecheck clean (2173 files), agent tests 9/9 green.
